Docs folder standardization and other fixes
Differences from the testing.md file have been added to the internals.md file

+++ src/ColumnSchema.php
@@ -1,8 +1,9 @@
 <?php
+
 /**
- * @link http://www.yiiframework.com/
- * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
+ * @link https://www.yiiframework.com/
+ * @copyright Copyright © 2008 by Yii Software (https://www.yiiframework.com/)
+ * @license https://www.yiiframework.com/license/
  */
 
 namespace Yiisoft\Db\Sphinx;
@@ -106,7 +107,7 @@ class ColumnSchema extends BaseObject
             case 'boolean':
                 return (bool) $value;
             case 'double':
-                return (double) $value;
+                return (float) $value;
         }
 
         return $value;
